requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.3
scikit-learn==1.0.2
pandas==1.4.2
numpy==1.22.3
Werkzeug==2.2.3
SQLAlchemy==1.4.46